move job model to separate file
